Add test for no results (i.e. if first iteration)
This could be better implemented using the GetListOfPortMappings
api. However, it does require a second parsing of the returned
CDATA, and the returned object is of a different format.

For miniupnpd, the current implementation will return an error
with the text `ArrayIndexInvalid` - so on the first iteration, we
can use this to determine if there are no records available, instead
of failing on the next test and returning an error - when there isn't.

If there is another upnp implementation being used, it will fall back
to the prior behavior.